Groundwork for rare weapon loot, used for LegenDoom compatibility
This will soon be used to allow players to find higher-grade weapons from level pre-placements or enemy drops if they are already carrying weapons at the corresponding grade. However, for now this feature is on the backburner, since I'd like to be able to put at least one weapon type in each category for each grade before implementing this feature in full.