Better auto-reload code
`BIO_Weapon::A_AutoReload()` had a bug preventing correct calculation of minimum ammo needed to not reload, and there needed to be a separate function for auto-reloading weapons after firing if their magazine had just been emptied.

Additionally, the new pre-fire ammo checker looks for a state labelled 'Dryfire` and goes to it if both firing and reloading are impossible, either for lack of ammo or the rule of a CVar.

This commit also fixes the possibility of firing 4 barrels of the Volley Gun when only two shells are loaded.