Better randomising mutagen
The number of affixes generated is now narrowed after being initially generated by regenerating it at random from a range between the floor of `Min(2, BIO_Weapon::MaxAffixes)` and itself. This turns the number of affixes generated on a weapon from a flat distribution to a steep curve with ~50% of a sample of 10000 leading to 2 affixes, and only 4% leading to 6.

Also, prevent use of a randomiser mutagen with a weapon that has `MaxAffixes` less than 1, put an error message if a caller tries to randomise a weapon like this anyway, don't force the affix count floor to 2 even if the weapon's max affixes is 1, and create a debug mutagen for testing the generation algorithm.