Support defining weapon upgrade recipe templates
A token can now be given to either the input or output field which specifies a grade and weapon category. The autogen pool corresponding to these parameters will then be used to generate recipes against everything in the opposing field.

This way, for example, a weapon can be excluded from full autogeneration but be made into an input for all other weapons in the same grade and category.