New ammo scavenging system
I wasn't satisfied with the old way this was done so it's been revamped to work more like Zhs2's Intelligent Supplies.

- All weapons have some ammo to give and it gets siphoned off of them as you walk past, regardless of their rarity or grade.
- The weapon is only destroyed if it's been emptied, is standard-grade, is common rarity, and drops frequently, such as the Shotgun, Chaingun, or Pistol.
- The max-scavenge-grade CVar has been removed.
- Weapons now give however much ammo they can when attaching to a new owner.
- There's now a special pickup message when touching a weapon only grants ammo and not the weapon itself.

This does mean that Intelligent Supplies will readily swallow up rare/non-standard weapons if used out-of-the-box, but this is not enough to outweight the benefits of the new system. A different recourse will have to be sought out.