Begin pseudo-merge of branch `redux`
I'm certain that Biomorph needs to change from what's currently on the master branch. That incarnation makes several design decisions that I now find disagreeable.

Some of the features I want out of the new design are certain, such as waste-proof health, armour, and ammo items, but for the most part I do not yet have a clear vision as to what the gameplay of the "redux" will be.

As such, I'm making the call to purge the master branch and perform something akin to a merge, by selectively copying files from the branch to trunk until I have a clean starting point off of which to create multiple new topic branches, on which I can experiment with different gameplay schemes, rather than shuffling one branch's codebase over and over.

As such, the repository now contains only:
- A much-truncated attributions document.
- The updated .gitignore.
- A generalised README.
- The Blender file.
- The player class patch.