Begin branch `genegraph`
This recycles much of the code from the random weapon upgrade graph experiment to create a customisable graph instead. This solves virtually all of the problems with that aforementioned design, and will probably lead to much more fun and emergence in the long run.

This codebase still has a long way to go before it is serviceable for playtesting, but the gist is there.

Details:
- Weapon and payload code.
- 1-to-50 chance of a random pistol spawning alongside a `Clip` item.
- The modifier graph class and its representative menu.
- Some simple modifiers, including a damage addition, magazine size increaser, infinite ammo, and demonslayer (LegenDoom).
- Mutagens, genes, related GLDEFS.
- /docs/code-tour.md.
- Sounds and SNDINFO.
- GIMP files for the graph node graphic, mutagen sprites, and gene sprites.