Gives genes a loot table, and start spawning them
Note that this changes the loot value balance dramatically; the threshold for something to drop has been reduced from 3200 to 600. The next commit will change how general mutagens are spent to bring back equilibrium.

The loot weights are a work-in-progress, since it'll be impossible to balance them properly until the pool is more saturated with a variety of choices.