Fancify weapon mod menu node connection rendering
The previous style left little distinction besides the drawn colour. A trivial solution may have been to thicken the lines but I thought I would take this opportunity to add some suitable flavour, with the new lines being made of segments that pulse, which suits the organic themes of the mod as a whole.