Tweak and nerf the Demonslayer weapon modifier
This was much too powerful, firstly because it used the same 400% as an effect which, in its own mod, assumes it's the only effect on the weapon. On puffs, it also damages the monster for 300% of the original damage, then multiplied by 4 because of the damage type, leading to massive overtuning.

Don't use the `DemonSlayer` damage type anymore, and reduce the benefit of the modifier to a 100% increase.