Rework how vanilla weapons get replaced
The previous solution left certain DEHACKED pickups untouched and able to give a player a non-Biomorph weapon. The patch that had been implemented in prototype 2 was to use an inventory item attached to the pawn that would intercept these pickups and quickly replace them, but this solution was flawed, since the player saw a vanilla pickup message and couldn't know what weapon they would get until they walked on an out-of-place-looking vanilla item.

Now, use the `replaces` keyword to replace all vanilla weapons with "markers" that then get spawn-handled to be replaced with Biomorph weapons.