Replace `BIO_PayloadCommon::MetaFlags` with `SizeClass`
For starters, setting the `TubeBased` bit on `BIO_Rocket` caused it to ride along the ground regardless of the user's aim pitch, which reeks of a memory error. Meta flagdefs have also proven a crapshoot, since the compiler permits the bits on an instance to be modified at runtime, bizarrely.

The new enumeration is better at broadly communicating intent, to boot.