Refactor payload functor system
Instead of allocating 2-3 dynamic arrays for every single payload fired, keep a pointer to an object "owned" by the pipeline, and iterate over the arrays in that.

Additionally, remove `BIO_Projectile::Acceleration` and `BIO_Projectile::SeekAngle`. This functionality belongs to travel functors, and will be provided as such later on through modifier genes.