Expand conditions for morphing a weapon into its mirrored variant
The selection of avenues for improvement the user needs to draw from to be eligible has increased, so more different combinations of genes work, although now the user has to fill the graph with one different improvement in each node. However, now the user can also fill the graph entirely with one type of gene.

Note that the logic for the former condition is untested since there aren't enough different genes in the pool yet to actually make the Machine Gun eligible for a morph.