Gene/modifier data model overhaul
Genes are now stateful, with an object that gets moved from a gene into the mod graph and then can be moved out again, back into a gene item. This enables genes with any number of modifiers, and allows those modifiers to be assigned procedurally.

The high-level design goal of this change is to encourage more weapon customisation by pairing improvements with detriments, leading players to experiment with modifier combinations to minimise disadvantages and maximise advantages.

Additionally, support and active effects are getting merged into modifiers, so that a gene can have a support function and a weapon-modifying function simultaneously, for instance.

Finally:
- Remove `BIO_GeneContext::TotalCount` and `First`, since they weren't being used.
- `BIO_Gene::LockOnCommit` is gone.
- Remove text colours `MGene` and `SGene`. Rename `AGene` to `Pink` since ZDoom has nothing of the sort natively, and it may be useful.