Remove Minivulcan, Turbovulcan
Biomorph is henceforth going to lean towards creating more variance out of a smaller set of base weapons. When "flavour rules" are implemented, maybe these two names can be assigned to Microvulcans that meet a certain set of conditions. The same goes for the sounds. Those files, though their SNDINFO entries are being purged, are the only things to remain of these two weapons.