Remove Pump Shotgun, Auto Shotgun, Coachgun
An upcoming commit will create a new, more generalised magazine-fed pump-action shotgun with the option of being modified to become fully automatic. The Auto Shotgun's sprites will be used as placeholders for a new Super Shotgun replacement. In between then and now, however, operating modes are going to be removed, so let's reduce the amount of work that will take. I'm still debating removing the Volley Gun.