New Shotgun and Stormgun
The former is an 8-round magazine-fed pump-action, while the latter is akin to Final Doomer's Shattergun, or DRLA's Shredder Shotgun.

The intent here is to reduce the weapon count even further than it already is, so the Auto Shotgun and Pump Shotgun are essentially going to be merged, with the new Shotgun eventually having a modifier that removes the need to pump it.

The Stormgun is there because I found myself uninspired by the idea of an over-under double-barreled shotgun, when I could go with a superposed-load weapon, which I could then model visually after Crysis 3's Typhoon. It's a weapon that I think offers deeper gameplay and more opportunities for interesting modification.