Moving towards the tentative final new gameplay design
The plan is to have a more DRLA-like system where the player gets 8 slots with which they may do what they please (even if that means, for instance, having 8 different kinds of pistol). This will enable more freedom in customising one's arsenal; now all that's left to do is work out implementation details.

This commit is a step in the right direction, although to really have a vertical slice will require the menu that powers it all, which I would like to build off of ZForms after it gets updated.