Only register player's default super shotgun conditionally
If the player is using an IWAD that isn't from the "Doom 2 family" an SSG never becomes a part of their arsenal. This should make no functional difference unless another mod somehow spawns an SSG pickup but this seems like a more correct thing to do from a "user expectations" standpoint.