New player class: "Lapsing Biomorph"
The pistol-start playstyle made manifest. No need for a third-party mod to reset the inventory every map, better for discoverability than CVars (or at least it will be when the menu actually tells players what this does), and this opens the way to changing the gameplay for users who want to play this way.