Reduce range of unarmed attacks
The original value was pulled straight from Final Doomer's Katana because I wanted to see how balanced it would prove to be in this context. My conclusions are that it doesn't feel right given the animations and that it would much better if one of the advantages of keeping a melee weapon registered in one's arsenal was a significant range increase over unarmed attacks.